feat: Rework blogs page to eye health page
This commit refactors the blogs page to become the eye health page and updates relevant links and titles.

- Changed the "View all Blogs" link in `src/sections/HomePage/Blogs.tsx` to point to `/eye-health`.
- Added an "Eye Health" link to the footer in `src/sections/HomePage/Footer.tsx`.
- Updated the "Eye Health" link in the header in `src/sections/HomePage/Header.tsx` to point to `/eye-health`.
- Replaced the contents of `src/app/blogs/page.tsx` with the "coming soon" page for eye health.
- Deleted the original "coming soon" page at `src/app/eyehealth/page.tsx`.
